ai/pyproject.toml
[project]
name = "bisleri-ai"
version = "0.1.0"
description = "AI services for rural women entrepreneurship marketplace"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "ultralytics>=8.3.0",
    "Pillow>=10.0.0",
    "openai>=1.58.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "selectolax>=0.3.27",
    "fake-useragent>=2.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "kaggle>=1.6.0",
]
